requirements.txt
chromadb
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
openai
python-dotenv
scipy
sentence-transformers
streamlit
